// types/companyDataType.ts
export interface CompanyDataType {
  name: string;
  phoneNumber: string;
  businessIdentification?: string;
  companyEstablish?: string;
  companyName?: string;
  companyWebsite?: string;
  countryName?: string;
  numberOfEmployees?: string | number;
  businessCategory: string;
  businessType: any;
  subCategory: string;
  about: string;
  mainProducts: string;
  companyAddress: string;
  zipCode: string;
  district: string;
  fullName: string;
  isVerified: boolean;
  ownerName: string;
  ownerEmail: string;
  ownerPhone: string;
  tradeLicense: string;
}